chore(Kaltura Player API): moving code from VideoPlayerAPI to the deprecation of it
